feat: add iOS NativeTrackingPlugin for background location tracking and audio keep-alive

+++ ios/App/App/NativeTrackingPlugin.swift
@@ -262,9 +262,9 @@ public class NativeTrackingPlugin: CAPPlugin, CLLocationManagerDelegate {
         data.append(Data(bytes: &chunkSizeLE, count: 4))
         data.append(contentsOf: [0x57, 0x41, 0x56, 0x45]) // "WAVE"
         data.append(contentsOf: [0x66, 0x6D, 0x74, 0x20]) // "fmt "
-        var subchunk1Size: Int32 = 16.littleEndian
+        var subchunk1Size: Int32 = Int32(16).littleEndian
         data.append(Data(bytes: &subchunk1Size, count: 4))
-        var audioFormat: Int16 = 1.littleEndian // PCM
+        var audioFormat: Int16 = Int16(1).littleEndian // PCM
         data.append(Data(bytes: &audioFormat, count: 2))
         var channelsLE = numChannels.littleEndian
         data.append(Data(bytes: &channelsLE, count: 2))
